pyproject.toml
[tool.poetry]
name = "rag-semantic-search"
version = "0.1.0"
description = "RAG Semantic Search Pipeline with Qwen Embeddings and Qdrant"
authors = ["hoangmai"]

[tool.poetry.dependencies]
python = "^3.12"
gradio = "^5.0.0"
qdrant-client = "^1.6.0"
transformers = "^4.35.0"
torch = "^2.1.0"
openai = "^1.3.0"
pypdf = "^3.17.0"
python-docx = "^1.1.0"
sentence-transformers = "^2.2.2"
numpy = "^1.24.0"
pandas = "^2.1.0"
python-dotenv = "^1.0.0"
tiktoken = "^0.5.0"
nltk = "^3.8.1"
# New dependencies for Nanonets OCR integration
pdf2image = "^1.16.0"
pillow = "^10.0.0"
beautifulsoup4 = "^4.12.0"
accelerate = "^1.9.0"
einops = "^0.7.0"
timm = "^0.9.0"
sentencepiece = "^0.1.99"
protobuf = "^4.25.0"
bitsandbytes = "^0.41.0"
docling = "^2.41.0"
langchain = "^0.3.26"
natsort = "^8.4.0"
pymupdf = "^1.26.3"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.0"
black = "^23.9.0"
flake8 = "^6.1.0"
mypy = "^1.6.0"
ipykernel = "^6.29.5"
jupyter = "^1.1.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api" 